requirements.txt
python-telegram-bot
PyMySQL